\name{Generate.Powers}
\alias{Generate.Powers}

\title{Generating Powers Series of Variables}
\description{
Internal function used by \link{npse} to generate covariates power series.
}

\usage{
Generate.Powers(X, lambda)
}

\arguments{
  \item{X}{covariates.}
  \item{lambda}{the maximal order of power series.}
}

\author{
Weihua An, Departments of Statistics and Sociology, Indiana University Bloomington, \email{weihuaan@indiana.edu}. \cr
Xuefu Wang, Department of Statistics, Indiana University Bloomington, \email{wangxuef@umail.iu.edu}.
}

\seealso{
\code{\link{larf}}, \code{\link{npse}}
}

