\name{print.larf}
\alias{print.larf}


\title{Print Results of the Estimated LARF}
\description{
  Methods to display brief results of a \code{\link{larf}} object.
}

\usage{
\method{print}{larf}(x, digits = 4, \dots)
}

\arguments{
  \item{x}{an object of class \code{"larf"} as fitted by \code{\link{larf}}.}
  \item{digits}{The number of significant digits to be printed in the reports of the results.}
  \item{\dots}{currently not used.}
}

\author{
Weihua An and Xuefu Wang, Departments of Statistics and Sociology, Indiana University Bloomington
}

\seealso{
\code{\link{larf}}, \code{\link{larf.fit}}
}
