\name{summary.larf}
\alias{summary.larf}
\alias{print.summary.larf}


\title{Summary of the Estimated LARF}
\description{
  Summary of an object in the \code{\link{larf}} class.
}

\usage{
\method{summary}{larf}(object, \dots)
}

\arguments{
  \item{object}{an object of class \code{"larf"} as fitted by \code{\link{larf}}.}
  \item{\dots}{currently not used.}
}

\author{
Weihua An and Xuefu Wang, Departments of Statistics and Sociology, Indiana University Bloomington
}

\seealso{
\code{\link{larf}}, \code{\link{larf.fit}}
}

