\name{vcov.larf}
\alias{vcov.larf}


\title{Local Average Response Functions}
\description{
  Methods to display the variance covariance matrix of the estimated parameters for the regression fitted by \code{\link{larf}}.
}

\usage{
\method{vcov}{larf}(object, \dots)
}

\arguments{
  \item{object}{an object of class \code{"larf"} as fitted by \code{\link{larf}}.}
  \item{\dots}{currently not used.}
}


\seealso{
\code{\link{larf}}, \code{\link{larf.fit}}
}
