\name{print.larf}
\alias{print.larf}


\title{Local Average Response Functions}
\description{
  Methods to display brief results of the instrumental-variable regressions fitted by \code{\link{larf}}.
}

\usage{
\method{print}{larf}(x, digits = 3, \dots)
}

\arguments{
  \item{x}{an object of class \code{"larf"} as fitted by \code{\link{larf}}.}
  \item{digits}{The desired minimum number of significant digits to be printed in the reports of the results. The default is 3.}
  \item{\dots}{currently not used.}
}


\seealso{
\code{\link{larf}}, \code{\link{larf.fit}}
}
