\name{vcov.larf}
\alias{vcov.larf}


\title{Variance Covariance Matrix of the Parameters in the Estimated LARF}
\description{
  Methods to display the variance covariance matrix of the model parameters estimated by \code{\link{larf}}.
}

\usage{
\method{vcov}{larf}(object, \dots)
}

\arguments{
  \item{object}{an object of class \code{"larf"} as fitted by \code{\link{larf}}.}
  \item{\dots}{currently not used.}
}

\author{
Weihua An and Xuefu Wang, Departments of Statistics and Sociology, Indiana University Bloomington
}

\seealso{
\code{\link{larf}}, \code{\link{larf.fit}}
}
